% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-solutions.R
\docType{data}
\name{aoc18_solutions}
\alias{aoc18_solutions}
\title{Advent of Code 2018 solutions}
\format{An object of class \code{list} of length 18.}
\usage{
aoc18_solutions
}
\description{
Advent of Code 2018 solutions
}
\keyword{datasets}
