% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{day08}
\alias{day08}
\alias{parse_license}
\alias{parse_license2}
\title{Day 08: Memory Maneuver}
\usage{
parse_license(x)

parse_license2(x)
}
\arguments{
\item{x}{a license (a single-string with space-separated numbers)}
}
\value{
For Part One, \code{parse_license(x)} returns a vector of metadata values.
For Part Two, \code{parse_license2(x)} returns the value of the root node.
}
\description{
\href{https://adventofcode.com/2018/day/8}{Memory Maneuver}
}
\details{
\strong{Part One}

The sleigh is much easier to pull than you'd expect for something its
weight. Unfortunately, neither you nor the Elves know [which
way]{title="It's North. Obviously."} the North Pole is from here.

You check your wrist device for anything that might help. It seems to
have some kind of navigation system! Activating the navigation system
produces more bad news: "Failed to start navigation system. Could not
read software license file."

The navigation system's license file consists of a list of numbers
(your puzzle input). The numbers define a data structure which, when
processed, produces some kind of
\href{https://en.wikipedia.org/wiki/Tree_(data_structure)}{tree} that can be
used to calculate the license number.

The \emph{tree} is made up of \emph{nodes}; a single, outermost node forms the
tree's \emph{root}, and it contains all other nodes in the tree (or contains
nodes that contain nodes, and so on).

Specifically, a node consists of:
\itemize{
\item A \emph{header}, which is always exactly two numbers:
\itemize{
\item The quantity of child nodes.
\item The quantity of metadata entries.
}
\item Zero or more \emph{child nodes} (as specified in the header).
\item One or more \emph{metadata entries} (as specified in the header).
}

Each child node is itself a node that has its own header, child nodes,
and metadata. For example:\preformatted{2 3 0 3 10 11 12 1 1 0 1 99 2 1 1 2
A----------------------------------
    B----------- C-----------
                     D-----
}

In this example, each node of the tree is also marked with an underline
starting with a letter for easier identification. In it, there are four
nodes:
\itemize{
\item \code{A}, which has \code{2} child nodes (\code{B}, \code{C}) and \code{3} metadata entries
(\code{1}, \code{1}, \code{2}).
\item \code{B}, which has \code{0} child nodes and \code{3} metadata entries (\code{10}, \code{11},
\code{12}).
\item \code{C}, which has \code{1} child node (\code{D}) and \code{1} metadata entry (\code{2}).
\item \code{D}, which has \code{0} child nodes and \code{1} metadata entry (\code{99}).
}

The first check done on the license file is to simply add up all of the
metadata entries. In this example, that sum is
\code{1+1+2+10+11+12+2+99=138}.

\emph{What is the sum of all metadata entries?}

\strong{Part Two}

The second check is slightly more complicated: you need to find the
value of the root node (\code{A} in the example above).

The \emph{value of a node} depends on whether it has child nodes.

If a node has \emph{no child nodes}, its value is the sum of its metadata
entries. So, the value of node \code{B} is \code{10+11+12=33}, and the value of
node \code{D} is \code{99}.

However, if a node \emph{does have child nodes}, the metadata entries become
indexes which refer to those child nodes. A metadata entry of \code{1} refers
to the first child node, \code{2} to the second, \code{3} to the third, and so on.
The value of this node is the sum of the values of the child nodes
referenced by the metadata entries. If a referenced child node does not
exist, that reference is skipped. A child node can be referenced
multiple time and counts each time it is referenced. A metadata entry of
\code{0} does not refer to any child node.

For example, again using the above nodes:
\itemize{
\item Node \code{C} has one metadata entry, \code{2}. Because node \code{C} has only one
child node, \code{2} references a child node which does not exist, and so
the value of node \code{C} is \code{0}.
\item Node \code{A} has three metadata entries: \code{1}, \code{1}, and \code{2}. The \code{1}
references node \code{A}'s first child node, \code{B}, and the \code{2} references
node \code{A}'s second child node, \code{C}. Because node B has a value of
\code{33} and node \code{C} has a value of \code{0}, the value of node \code{A} is
\code{33+33+0=66}.
}

So, in this example, the value of the root node is \code{66}.

\emph{What is the value of the root node?}
}
\examples{
x <- "2 3 0 3 10 11 12 1 1 0 1 99 2 1 1 2"
parse_license(x)
parse_license2(x)
}
