% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{day09}
\alias{day09}
\alias{get_marble_high_score}
\alias{run_marbles}
\alias{parse_marble_description}
\alias{f09b}
\title{Day 09: Marble Mania}
\usage{
get_marble_high_score(marble_history)

run_marbles(players, marbles)

parse_marble_description(x)

f09b(x)
}
\arguments{
\item{x}{some data}
}
\value{
For Part One, \code{f09a(x)} returns .... For Part Two,
\code{f09b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2018/day/9}{Marble Mania}
}
\details{
\strong{Part One}

You talk to the Elves while you wait for your navigation system to
[initialize]{title="Do you have any idea how long it takes to load navigation data for all of time and space?!"}.
To pass the time, they introduce you to their favorite
\href{https://en.wikipedia.org/wiki/Marble_(toy)}{marble} game.

The Elves play this game by taking turns arranging the marbles in a
\emph{circle} according to very particular rules. The marbles are numbered
starting with \code{0} and increasing by \code{1} until every marble has a number.

First, the marble numbered \code{0} is placed in the circle. At this point,
while it contains only a single marble, it is still a circle: the marble
is both clockwise from itself and counter-clockwise from itself. This
marble is designated the \emph{current marble}.

Then, each Elf takes a turn placing the \emph{lowest-numbered remaining
marble} into the circle between the marbles that are \code{1} and \code{2} marbles
\emph{clockwise} of the current marble. (When the circle is large enough,
this means that there is one marble between the marble that was just
placed and the current marble.) The marble that was just placed then
becomes the \emph{current marble}.

However, if the marble that is about to be placed has a number which is
a multiple of \code{23}, \emph{something entirely different happens}. First, the
current player keeps the marble they would have placed, adding it to
their \emph{score}. In addition, the marble \code{7} marbles \emph{counter-clockwise}
from the current marble is \emph{removed} from the circle and \emph{also} added to
the current player's score. The marble located immediately \emph{clockwise}
of the marble that was removed becomes the new \emph{current marble}.

For example, suppose there are 9 players. After the marble with value
\code{0} is placed in the middle, each player (shown in square brackets)
takes a turn. The result of each of those turns would produce circles of
marbles like this, where clockwise is to the right and the resulting
current marble is in parentheses:\preformatted{[-] (0)
[1]  0 (1)
[2]  0 (2) 1
[3]  0  2  1 (3)
[4]  0 (4) 2  1  3
[5]  0  4  2 (5) 1  3
[6]  0  4  2  5  1 (6) 3
[7]  0  4  2  5  1  6  3 (7)
[8]  0 (8) 4  2  5  1  6  3  7
[9]  0  8  4 (9) 2  5  1  6  3  7
[1]  0  8  4  9  2(10) 5  1  6  3  7
[2]  0  8  4  9  2 10  5(11) 1  6  3  7
[3]  0  8  4  9  2 10  5 11  1(12) 6  3  7
[4]  0  8  4  9  2 10  5 11  1 12  6(13) 3  7
[5]  0  8  4  9  2 10  5 11  1 12  6 13  3(14) 7
[6]  0  8  4  9  2 10  5 11  1 12  6 13  3 14  7(15)
[7]  0(16) 8  4  9  2 10  5 11  1 12  6 13  3 14  7 15
[8]  0 16  8(17) 4  9  2 10  5 11  1 12  6 13  3 14  7 15
[9]  0 16  8 17  4(18) 9  2 10  5 11  1 12  6 13  3 14  7 15
[1]  0 16  8 17  4 18  9(19) 2 10  5 11  1 12  6 13  3 14  7 15
[2]  0 16  8 17  4 18  9 19  2(20)10  5 11  1 12  6 13  3 14  7 15
[3]  0 16  8 17  4 18  9 19  2 20 10(21) 5 11  1 12  6 13  3 14  7 15
[4]  0 16  8 17  4 18  9 19  2 20 10 21  5(22)11  1 12  6 13  3 14  7 15
[5]  0 16  8 17  4 18(19) 2 20 10 21  5 22 11  1 12  6 13  3 14  7 15
[6]  0 16  8 17  4 18 19  2(24)20 10 21  5 22 11  1 12  6 13  3 14  7 15
[7]  0 16  8 17  4 18 19  2 24 20(25)10 21  5 22 11  1 12  6 13  3 14  7 15
}

The goal is to be the \emph{player with the highest score} after the last
marble is used up. Assuming the example above ends after the marble
numbered \code{25}, the winning score is \code{23+9=32} (because player 5 kept
marble \code{23} and removed marble \code{9}, while no other player got any points
in this very short example game).

Here are a few more examples:
\itemize{
\item \code{10} players; last marble is worth \code{1618} points: high score is
\emph{\code{8317}}
\item \code{13} players; last marble is worth \code{7999} points: high score is
\emph{\code{146373}}
\item \code{17} players; last marble is worth \code{1104} points: high score is
\emph{\code{2764}}
\item \code{21} players; last marble is worth \code{6111} points: high score is
\emph{\code{54718}}
\item \code{30} players; last marble is worth \code{5807} points: high score is
\emph{\code{37305}}
}

\emph{What is the winning Elf's score?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f09a()
f09b()
}
