% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day11.R
\name{day11}
\alias{day11}
\alias{pool_power_cells}
\alias{compute_cell_power_level}
\alias{create_power_cell_matrix}
\alias{search_power_cell_pools}
\title{Day 11: Chronal Charge}
\usage{
pool_power_cells(mat, pool_span, fun = sum)

compute_cell_power_level(x, y, serial_number)

create_power_cell_matrix(span, serial_number)

search_power_cell_pools(mat, to_search = NULL)
}
\arguments{
\item{x}{some data}
}
\value{
For Part One, \code{f11a(x)} returns .... For Part Two,
\code{f11b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2018/day/11}{Chronal Charge}
}
\details{
\strong{Part One}

You watch the Elves and their sleigh fade into the distance as they head
toward the North Pole.

Actually, you're the one fading. The [falling
sensation]{title="wheeeeeeeeeeeeeeeeee"} returns.

The low fuel warning light is illuminated on your wrist-mounted device.
Tapping it once causes it to project a hologram of the situation: a
\emph{300x300} grid of fuel cells and their current power levels, some
negative. You're not sure what negative power means in the context of
time travel, but it can't be good.

Each fuel cell has a coordinate ranging \emph{from 1 to 300} in both the X
(horizontal) and Y (vertical) direction. In \code{X,Y} notation, the top-left
cell is \code{1,1}, and the top-right cell is \code{300,1}.

The interface lets you select \emph{any 3x3 square} of fuel cells. To
increase your chances of getting to your destination, you decide to
choose the 3x3 square with the \emph{largest total power}.

The power level in a given fuel cell can be found through the following
process:
\itemize{
\item Find the fuel cell's \emph{rack ID}, which is its \emph{X coordinate plus 10}.
\item Begin with a power level of the \emph{rack ID} times the \emph{Y coordinate}.
\item Increase the power level by the value of the \emph{grid serial number}
(your puzzle input).
\item Set the power level to itself multiplied by the \emph{rack ID}.
\item Keep only the \emph{hundreds digit} of the power level (so \code{12345}
becomes \code{3}; numbers with no hundreds digit become \code{0}).
\item \emph{Subtract 5} from the power level.
}

For example, to find the power level of the fuel cell at \code{3,5} in a grid
with serial number \code{8}:
\itemize{
\item The rack ID is \code{3 + 10 = 13}.
\item The power level starts at \code{13 * 5 = 65}.
\item Adding the serial number produces \code{65 + 8 = 73}.
\item Multiplying by the rack ID produces \code{73 * 13 = 949}.
\item The hundreds digit of \code{949} is \code{9}.
\item Subtracting 5 produces \code{9 - 5 = 4}.
}

So, the power level of this fuel cell is \code{4}.

Here are some more example power levels:
\itemize{
\item Fuel cell at  \code{122,79}, grid serial number \code{57}: power level \code{-5}.
\item Fuel cell at \code{217,196}, grid serial number \code{39}: power level  \code{0}.
\item Fuel cell at \code{101,153}, grid serial number \code{71}: power level  \code{4}.
}

Your goal is to find the 3x3 square which has the largest total power.
The square must be entirely within the 300x300 grid. Identify this
square using the \code{X,Y} coordinate of its \emph{top-left fuel cell}. For
example:

For grid serial number \code{18}, the largest total 3x3 square has a top-left
corner of \code{33,45} (with a total power of \code{29}); these fuel cells appear
in the middle of this 5x5 region:\preformatted{-2  -4   4   4   4
-4   4   4   4  -5
 4   3   3   4  -4
 1   1   2   4  -3
-1   0   2  -5  -2
}

For grid serial number \code{42}, the largest 3x3 square's top-left is
\code{21,61} (with a total power of \code{30}); they are in the middle of this
region:\preformatted{-3   4   2   2   2
-4   4   3   3   4
-5   3   3   4  -4
 4   3   3   4  -3
 3   3   3  -5  -1
}

\emph{What is the \code{X,Y} coordinate of the top-left fuel cell of the 3x3
square with the largest total power?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f11a()
f11b()
}
