% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{day01}
\alias{day01}
\alias{f1}
\alias{f2}
\title{Day 01: Chronal Calibration}
\usage{
f1(x)

f2(x)
}
\arguments{
\item{x}{some data}
}
\value{
some values
}
\description{
\href{https://adventofcode.com/2018/day/1}{Chronal Calibration}
}
\details{
\strong{Part One}

"We've detected some temporal anomalies," one of Santa's Elves at the
[Temporal Anomaly Research and Detection Instrument
Station]{title="It's about as big on the inside as you expected."} tells
you. She sounded pretty worried when she called you down here. "At
500-year intervals into the past, someone has been changing Santa's
history!"

"The good news is that the changes won't propagate to our time stream
for another 25 days, and we have a device" - she attaches something to
your wrist - "that will let you fix the changes with no such propagation
delay. It's configured to send you 500 years further into the past every
few days; that was the best we could do on such short notice."

"The bad news is that we are detecting roughly \emph{fifty} anomalies
throughout time; the device will indicate fixed anomalies with \emph{stars}.
The other bad news is that we only have one device and you're the best
person for the job! Good lu--" She taps a button on the device and you
suddenly feel like you're falling. To save Christmas, you need to get
all \emph{fifty stars} by December 25th.

Collect stars by solving puzzles. Two puzzles will be made available on
each day in the advent calendar; the second puzzle is unlocked when you
complete the first. Each puzzle grants \emph{one star}. Good luck!

After feeling like you've been falling for a few minutes, you look at
the device's tiny screen. "Error: Device must be calibrated before first
use. Frequency drift detected. Cannot maintain destination lock." Below
the message, the device shows a sequence of changes in frequency (your
puzzle input). A value like \code{+6} means the current frequency increases
by \code{6}; a value like \code{-3} means the current frequency decreases by \code{3}.

For example, if the device displays frequency changes of
\code{+1, -2, +3, +1}, then starting from a frequency of zero, the following
changes would occur:
\itemize{
\item Current frequency \code{ 0}, change of \code{+1}; resulting frequency \code{ 1}.
\item Current frequency \code{ 1}, change of \code{-2}; resulting frequency \code{-1}.
\item Current frequency \code{-1}, change of \code{+3}; resulting frequency \code{ 2}.
\item Current frequency \code{ 2}, change of \code{+1}; resulting frequency \code{ 3}.
}

In this example, the resulting frequency is \code{3}.

Here are other example situations:
\itemize{
\item \code{+1, +1, +1} results in \code{ 3}
\item \code{+1, +1, -2} results in \code{ 0}
\item \code{-1, -2, -3} results in \code{-6}
}

Starting with a frequency of zero, \emph{what is the resulting frequency}
after all of the changes in frequency have been applied?

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}
}
\examples{
f1()
f2()
}
