% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03}
\alias{day03}
\alias{count_overlapping_fabric_claims}
\alias{find_unique_fabric_claim}
\title{Day 03: No Matter How You Slice It}
\usage{
count_overlapping_fabric_claims(x)

find_unique_fabric_claim(x)
}
\arguments{
\item{x}{a character vector of fabric claims}
}
\value{
For Part One, \code{count_overlapping_fabric_claims(x)} returns the total
number of squares that are claimed by more than one fabric claim. For Part
Two, \code{find_unique_fabric_claim(x)} returns the ID of the claim that does
not overlap any others.
}
\description{
\href{https://adventofcode.com/2018/day/3}{No Matter How You Slice It}
}
\details{
\strong{Part One}

The Elves managed to locate the chimney-squeeze prototype fabric for
Santa's suit (thanks to [someone]{title="WAS IT YOU"} who helpfully
wrote its box IDs on the wall of the warehouse in the middle of the
night). Unfortunately, anomalies are still affecting them - nobody can
even agree on how to \emph{cut} the fabric.

The whole piece of fabric they're working on is a very large square - at
least \code{1000} inches on each side.

Each Elf has made a \emph{claim} about which area of fabric would be ideal
for Santa's suit. All claims have an ID and consist of a single
rectangle with edges parallel to the edges of the fabric. Each claim's
rectangle is defined as follows:
\itemize{
\item The number of inches between the left edge of the fabric and the
left edge of the rectangle.
\item The number of inches between the top edge of the fabric and the top
edge of the rectangle.
\item The width of the rectangle in inches.
\item The height of the rectangle in inches.
}

A claim like \code{#123 @ 3,2: 5x4} means that claim ID \code{123} specifies a
rectangle \code{3} inches from the left edge, \code{2} inches from the top edge,
\code{5} inches wide, and \code{4} inches tall. Visually, it claims the square
inches of fabric represented by \code{#} (and ignores the square inches of
fabric represented by \code{.}) in the diagram below:\preformatted{...........
...........
...#####...
...#####...
...#####...
...#####...
...........
...........
...........
}

The problem is that many of the claims \emph{overlap}, causing two or more
claims to cover part of the same areas. For example, consider the
following claims:\preformatted{#1 @ 1,3: 4x4
#2 @ 3,1: 4x4
#3 @ 5,5: 2x2
}

Visually, these claim the following areas:\preformatted{........
...2222.
...2222.
.11XX22.
.11XX22.
.111133.
.111133.
........
}

The four square inches marked with \code{X} are claimed by \emph{both \code{1} and
\code{2}}. (Claim \code{3}, while adjacent to the others, does not overlap either
of them.)

If the Elves all proceed with their own plans, none of them will have
enough fabric. \emph{How many square inches of fabric are within two or more
claims?}

\strong{Part Two}

Amidst the chaos, you notice that exactly one claim doesn't overlap by
even a single square inch of fabric with any other claim. If you can
somehow draw attention to it, maybe the Elves will be able to make
Santa's suit after all!

For example, in the claims above, only claim \code{3} is intact after all
claims are made.

\emph{What is the ID of the only claim that doesn't overlap?}
}
\examples{
x <- c(
  "#1 @ 1,3: 4x4",
  "#2 @ 3,1: 4x4",
  "#3 @ 5,5: 2x2")
count_overlapping_fabric_claims(x)
find_unique_fabric_claim(x)
}
